(web) ♻️ refactor: integrate profile page with server
