+++ apps/web/src/ui/profile/widgets/pages/Profile/Account/AdornmentsGroup/Adornment/index.tsx
@@ -1,21 +1,3 @@
-'use client'
+import { AdornmentView } from './AdornmentView'
 
-import Image from 'next/image'
-
-type AdornmentProps = {
-  title: string
-  image: string
-  value: string
-}
-
-export function Adornment({ title, image, value }: AdornmentProps) {
-  return (
-    <div className='flex flex-col items-center justify-center gap-3'>
-      <dt className='text-center text-green-500'>{title}</dt>
-      <div className='relative h-16 w-16'>
-        <Image src={image} fill alt='' />
-      </div>
-      <dd className='text-center text-sm text-gray-100'>{value}</dd>
-    </div>
-  )
-}
+export const Adornment = AdornmentView
