+++ apps/web/src/ui/profile/widgets/pages/Profile/ChallengesChart/Legend/index.tsx
@@ -1,21 +1,3 @@
-import { twMerge } from 'tailwind-merge'
+import { LegendView } from './LegendView'
 
-type LegendProps = {
-  label: string
-  value: number
-  total: number
-  color: string
-}
-
-export function Legend({ label, value, total, color }: LegendProps) {
-  return (
-    <div className='flex items-center justify-between gap-1'>
-      <span className={twMerge(color, 'mr-2 h-2 w-3 rounded-md')} />
-      <dt className='text-gray-300'>{label}</dt>
-      <dd className='text-lg font-semibold text-gray-100'>
-        {value}
-        <span className='text-base text-gray-500'>/{total}</span>
-      </dd>
-    </div>
-  )
-}
+export const Legend = LegendView
