+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/ChallengesListTab/index.tsx
@@ -1,56 +1,26 @@
-import { Datetime } from '@stardust/core/global/libs'
-
-import { ROUTES } from '@/constants'
-import { ShowMoreButton } from '@/ui/global/widgets/components/ShowMoreButton'
-import type { TabListSorter } from '../../TabListSorter'
+import { useRest } from '@/ui/global/hooks/useRest'
+import { ChallengesListTabView } from './ChallengesListTabView'
 import { useChallengesListTab } from './useChallengesListTab'
-import { RowsSkeleton } from '../RowsSkeleton'
-import * as Row from '../Row'
-import * as CraftMetrics from '../CraftMetrics'
-import { NoRowsMessage } from '../NoRowsMessage'
+import type { TabListSorter } from '../../TabListSorter'
 
-type ChallengesListProps = {
+type Props = {
   userId: string
   tabListSorter: TabListSorter
 }
 
-export function ChallengesListTab({ tabListSorter, userId }: ChallengesListProps) {
-  const { challenges, isRecheadedEnd, isLoading, nextPage } = useChallengesListTab(
+export const ChallengesListTab = ({ userId, tabListSorter }: Props) => {
+  const { challengingService } = useRest()
+  const { challenges, isLoading, isRecheadedEnd, nextPage } = useChallengesListTab(
+    challengingService,
     tabListSorter,
     userId,
   )
-
-  if (isLoading) {
-    return <RowsSkeleton />
-  }
-
-  if (!challenges.length) {
-    return <NoRowsMessage>Você ainda não criou nenhum desafio ainda.</NoRowsMessage>
-  }
-
   return (
-    <ul>
-      {challenges.map((challenge, index) => (
-        <Row.Container
-          key={challenge.id.value}
-          index={index}
-          href={ROUTES.challenging.challenges.challenge(challenge.slug.value)}
-          className='flex flex-col gap-2'
-        >
-          <Row.Title>{challenge.title.value}</Row.Title>
-          <CraftMetrics.Container>
-            <CraftMetrics.Metric icon='calendar'>
-              {new Datetime(challenge.postedAt).getRelativeTime()}
-            </CraftMetrics.Metric>
-            <CraftMetrics.Metric icon='upvote'>
-              {challenge.upvotesCount.value}
-            </CraftMetrics.Metric>
-          </CraftMetrics.Container>
-        </Row.Container>
-      ))}
-      {!isRecheadedEnd && (
-        <ShowMoreButton isLoading={isLoading} onClick={nextPage} className='mt-6' />
-      )}
-    </ul>
+    <ChallengesListTabView
+      challenges={challenges}
+      isLoading={isLoading}
+      isRecheadedEnd={isRecheadedEnd}
+      onShowMore={nextPage}
+    />
   )
 }
