+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/CraftMetrics/Container.tsx
@@ -1,6 +1,6 @@
 import type { ComponentProps } from 'react'
 
-export function Container({ children, ...divProps }: ComponentProps<'div'>) {
+export const Container = ({ children, ...divProps }: ComponentProps<'div'>) => {
   return (
     <div className='flex items-center gap-2' {...divProps}>
       {children}
