+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/NoRowsMessage/NoRowsMessageView.tsx
@@ -1,0 +1,17 @@
+import type { ComponentProps } from 'react'
+import { twMerge } from 'tailwind-merge'
+
+export const NoRowsMessageView = ({
+  children,
+  className,
+  ...pProps
+}: ComponentProps<'p'>) => {
+  return (
+    <p
+      className={twMerge('px-6 py-20 text-gray-500 text-center font-medium', className)}
+      {...pProps}
+    >
+      {children}
+    </p>
+  )
+}
