+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/NoRowsMessage/index.tsx
@@ -1,13 +1,3 @@
-import type { ComponentProps } from 'react'
-import { twMerge } from 'tailwind-merge'
+import { NoRowsMessageView } from './NoRowsMessageView'
 
-export function NoRowsMessage({ children, className, ...pProps }: ComponentProps<'p'>) {
-  return (
-    <p
-      className={twMerge('px-6 py-20 text-gray-500 text-center font-medium', className)}
-      {...pProps}
-    >
-      {children}
-    </p>
-  )
-}
+export const NoRowsMessage = NoRowsMessageView
