+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/RowsSkeleton/RowsSkeletonView.tsx
@@ -1,0 +1,10 @@
+export const RowsSkeletonView = () => {
+  return (
+    <ul>
+      <li className='w-full h-16 rounded-md bg-gray-700 animate-pulse' />
+      <li className='w-full h-16 rounded-md bg-gray-800 animate-pulse' />
+      <li className='w-full h-16 rounded-md bg-gray-700 animate-pulse' />
+      <li className='w-full h-16 rounded-md bg-gray-800 animate-pulse' />
+    </ul>
+  )
+}
