+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/RowsSkeleton/index.tsx
@@ -1,10 +1,3 @@
-export function RowsSkeleton() {
-  return (
-    <ul>
-      <li className='w-full h-16 rounded-md bg-gray-700 animate-pulse' />
-      <li className='w-full h-16 rounded-md bg-gray-800 animate-pulse' />
-      <li className='w-full h-16 rounded-md bg-gray-700 animate-pulse' />
-      <li className='w-full h-16 rounded-md bg-gray-800 animate-pulse' />
-    </ul>
-  )
-}
+import { RowsSkeletonView } from './RowsSkeletonView'
+
+export const RowsSkeleton = RowsSkeletonView
