+++ apps/web/src/ui/profile/widgets/pages/Profile/CraftsTable/TabsGroup/SolutionsListTab/index.tsx
@@ -1,59 +1,26 @@
-import { Datetime } from '@stardust/core/global/libs'
-
-import { ROUTES } from '@/constants'
-import { ShowMoreButton } from '@/ui/global/widgets/components/ShowMoreButton'
+import { useRest } from '@/ui/global/hooks/useRest'
 import type { TabListSorter } from '../../TabListSorter'
-import { NoRowsMessage } from '../NoRowsMessage'
 import { useSolutionsListTab } from './useSolutionsListTab'
-import { RowsSkeleton } from '../RowsSkeleton'
-import * as Row from '../Row'
-import * as CraftMetrics from '../CraftMetrics'
+import { SolutionsListTabView } from './SolutionsListTabView'
 
-type ChallengesListProps = {
+type Props = {
   userId: string
   tabListSorter: TabListSorter
 }
 
-export function SolutionsListTab({ tabListSorter, userId }: ChallengesListProps) {
+export function SolutionsListTab({ tabListSorter, userId }: Props) {
+  const { challengingService } = useRest()
   const { solutions, isRecheadedEnd, isLoading, nextPage } = useSolutionsListTab(
+    challengingService,
     tabListSorter,
     userId,
   )
-
-  if (isLoading) {
-    return <RowsSkeleton />
-  }
-
-  if (!solutions.length) {
-    return <NoRowsMessage>Você ainda não criou nenhuma solução ainda.</NoRowsMessage>
-  }
-
   return (
-    <ul>
-      {solutions.map((solution, index) => (
-        <Row.Container
-          key={solution.id.value}
-          index={index}
-          href={ROUTES.api.challenging.solution(solution.slug.value)}
-          className='flex flex-col gap-2'
-        >
-          <Row.Title>{solution.title.value}</Row.Title>
-          <CraftMetrics.Container>
-            <CraftMetrics.Metric icon='calendar'>
-              {new Datetime(solution.postedAt).getRelativeTime()}
-            </CraftMetrics.Metric>
-            <CraftMetrics.Metric icon='upvote'>
-              {solution.upvotesCount.value}
-            </CraftMetrics.Metric>
-            <CraftMetrics.Metric icon='eye'>
-              {solution.viewsCount.value}
-            </CraftMetrics.Metric>
-          </CraftMetrics.Container>
-        </Row.Container>
-      ))}
-      {!isRecheadedEnd && (
-        <ShowMoreButton isLoading={isLoading} onClick={nextPage} className='mt-3' />
-      )}
-    </ul>
+    <SolutionsListTabView
+      isLoading={isLoading}
+      solutions={solutions}
+      isRecheadedEnd={isRecheadedEnd}
+      onShowMore={nextPage}
+    />
   )
 }
