+++ apps/web/src/ui/profile/widgets/pages/Profile/UnlockedAchievementsList/EmptyListMessage/index.tsx
@@ -1,12 +1,3 @@
-'use client'
+import { EmptyListMessageView } from './EmptyListMessageView'
 
-import { Animation } from '@/ui/global/widgets/components/Animation'
-
-export function EmptyListMessage() {
-  return (
-    <div className='flex flex-col items-center justify-center'>
-      <Animation name='apollo-missing' size={220} hasLoop />
-      <p className='text-sm text-gray-300'>Nenhuma conquista adquirida ainda 😢.</p>
-    </div>
-  )
-}
+export const EmptyListMessage = EmptyListMessageView
